refactor: reservation예외 메서드명 변경
