--- src/main/java/roomescape/exception/InvalidRequestException.java
@@ -1,10 +1,0 @@
-package roomescape.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class InvalidRequestException extends ReservationException {
-
-    public InvalidRequestException(String message) {
-        super(message, HttpStatus.BAD_REQUEST, "INVALID_REQUEST");
-    }
-}
