--- src/main/java/roomescape/exception/NotFoundException.java
@@ -1,10 +1,0 @@
-package roomescape.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class NotFoundException extends ReservationException {
-
-    public NotFoundException(String message) {
-        super(message, HttpStatus.BAD_REQUEST, "RESOURCE_NOT_FOUND");
-    }
-}
